Refactored ReadOnlyTopicObjectFactory to ReadOnlyTopic
