+++ core/templates/domain/topic_viewer/read-only-topic.model.spec.ts
@@ -13,22 +13,16 @@
 // limitations under the License.
 
 /**
- * @fileoverview Tests for ReadOnlyTopicObjectFactory.
+ * @fileoverview Tests for ReadOnlyTopic.
  */
 
-import {
-  ReadOnlyTopic,
-  ReadOnlyTopicObjectFactory,
-} from 'domain/topic_viewer/read-only-topic-object.factory';
+import { ReadOnlyTopic } from 'domain/topic_viewer/read-only-topic.model';
 import {StoryNode} from 'domain/story/story-node.model';
 
 describe('Read only topic object Factory', () => {
-  let readOnlyTopicObjectFactory: ReadOnlyTopicObjectFactory;
   let _sampleReadOnlyTopic: ReadOnlyTopic;
 
   beforeEach(() => {
-    readOnlyTopicObjectFactory = new ReadOnlyTopicObjectFactory();
-
     let nodeDict = {
       id: 'node_1',
       thumbnail_filename: 'image.png',
@@ -106,7 +100,7 @@ describe('Read only topic object Factory', () => {
     };
 
     _sampleReadOnlyTopic =
-      readOnlyTopicObjectFactory.createFromBackendDict(sampleTopicDataDict);
+      ReadOnlyTopic.createFromBackendDict(sampleTopicDataDict);
   });
 
   it('should return correct values for read-only topic object', () => {
