+++ core/templates/domain/topic_viewer/topic-viewer-backend-api.service.spec.ts
@@ -24,9 +24,8 @@ import {TestBed, fakeAsync, flushMicrotasks} from '@angular/core/testing';
 
 import {
   ReadOnlyTopic,
-  ReadOnlyTopicBackendDict,
-  ReadOnlyTopicObjectFactory,
-} from 'domain/topic_viewer/read-only-topic-object.factory';
+  ReadOnlyTopicBackendDict
+} from 'domain/topic_viewer/read-only-topic.model';
 import {TopicViewerBackendApiService} from 'domain/topic_viewer/topic-viewer-backend-api.service';
 
 describe('Topic viewer backend API service', () => {
@@ -34,17 +33,13 @@ describe('Topic viewer backend API service', () => {
   let httpTestingController: HttpTestingController;
   let sampleDataResultsObjects: ReadOnlyTopic;
   let sampleDataResults: ReadOnlyTopicBackendDict;
-  let readOnlyTopicObjectFactory: ReadOnlyTopicObjectFactory;
 
   beforeEach(() => {
-    readOnlyTopicObjectFactory = new ReadOnlyTopicObjectFactory();
-
     TestBed.configureTestingModule({
       imports: [HttpClientTestingModule],
     });
     httpTestingController = TestBed.inject(HttpTestingController);
     topicViewerBackendApiService = TestBed.inject(TopicViewerBackendApiService);
-    readOnlyTopicObjectFactory = TestBed.inject(ReadOnlyTopicObjectFactory);
 
     let nodeDict = {
       id: 'node_1',
@@ -124,7 +119,7 @@ describe('Topic viewer backend API service', () => {
     };
 
     sampleDataResultsObjects =
-      readOnlyTopicObjectFactory.createFromBackendDict(sampleDataResults);
+      ReadOnlyTopic.createFromBackendDict(sampleDataResults);
   });
 
   afterEach(() => {
