+++ core/templates/pages/exploration-player-page/current-lesson-player/layout-directives/learner-view-info.component.spec.ts
@@ -31,8 +31,8 @@ import {
 } from '../../../../domain/exploration/read-only-exploration-backend-api.service';
 import {
   ReadOnlyTopicBackendDict,
-  ReadOnlyTopicObjectFactory,
-} from '../../../../domain/topic_viewer/read-only-topic-object.factory';
+  ReadOnlyTopic,
+} from 'domain/topic_viewer/read-only-topic.model';
 import {TopicViewerBackendApiService} from '../../../../domain/topic_viewer/topic-viewer-backend-api.service';
 import {UrlInterpolationService} from '../../../../domain/utilities/url-interpolation.service';
 import {PageContextService} from '../../../../services/page-context.service';
@@ -53,7 +53,6 @@ describe('Learner view info component', () => {
   let urlInterpolationService: UrlInterpolationService;
   let urlService: UrlService;
   let topicViewerBackendApiService: TopicViewerBackendApiService;
-  let readOnlyTopicObjectFactory: ReadOnlyTopicObjectFactory;
   let i18nLanguageCodeService: I18nLanguageCodeService;
 
   beforeEach(waitForAsync(() => {
@@ -68,7 +67,6 @@ describe('Learner view info component', () => {
         UrlInterpolationService,
         UrlService,
         TopicViewerBackendApiService,
-        ReadOnlyTopicObjectFactory,
       ],
       schemas: [NO_ERRORS_SCHEMA],
     }).compileComponents();
@@ -86,11 +84,10 @@ describe('Learner view info component', () => {
     urlInterpolationService = TestBed.inject(UrlInterpolationService);
     urlService = TestBed.inject(UrlService);
     i18nLanguageCodeService = TestBed.inject(I18nLanguageCodeService);
-    readOnlyTopicObjectFactory = TestBed.inject(ReadOnlyTopicObjectFactory);
     topicViewerBackendApiService = TestBed.inject(TopicViewerBackendApiService);
 
     spyOn(topicViewerBackendApiService, 'fetchTopicDataAsync').and.resolveTo(
-      readOnlyTopicObjectFactory.createFromBackendDict({
+      ReadOnlyTopic.createFromBackendDict({
         subtopics: [],
         skill_descriptions: {},
         uncategorized_skill_ids: [],
