+++ core/templates/pages/story-viewer-page/navbar-breadcrumb/story-viewer-navbar-breadcrumb.component.spec.ts
@@ -27,8 +27,8 @@ import {MockTranslatePipe} from 'tests/unit-test-utils';
 import {TopicViewerBackendApiService} from 'domain/topic_viewer/topic-viewer-backend-api.service';
 import {
   ReadOnlyTopicBackendDict,
-  ReadOnlyTopicObjectFactory,
-} from 'domain/topic_viewer/read-only-topic-object.factory';
+  ReadOnlyTopic,
+} from 'domain/topic_viewer/read-only-topic.model';
 import {I18nLanguageCodeService} from 'services/i18n-language-code.service';
 
 class MockUrlService {
@@ -47,7 +47,6 @@ class MockUrlService {
 
 let component: StoryViewerNavbarBreadcrumbComponent;
 let fixture: ComponentFixture<StoryViewerNavbarBreadcrumbComponent>;
-let readOnlyTopicObjectFactory: ReadOnlyTopicObjectFactory;
 let topicViewerBackendApiService: TopicViewerBackendApiService;
 let i18nLanguageCodeService: I18nLanguageCodeService;
 let urlService: UrlService;
@@ -83,7 +82,6 @@ describe('Subtopic viewer navbar breadcrumb component', () => {
   }));
 
   beforeEach(() => {
-    readOnlyTopicObjectFactory = TestBed.inject(ReadOnlyTopicObjectFactory);
     topicViewerBackendApiService = TestBed.inject(TopicViewerBackendApiService);
     i18nLanguageCodeService = TestBed.inject(I18nLanguageCodeService);
     urlService = TestBed.inject(UrlService);
@@ -92,7 +90,7 @@ describe('Subtopic viewer navbar breadcrumb component', () => {
     fixture.detectChanges();
 
     spyOn(topicViewerBackendApiService, 'fetchTopicDataAsync').and.resolveTo(
-      readOnlyTopicObjectFactory.createFromBackendDict({
+      ReadOnlyTopic.createFromBackendDict({
         subtopics: [],
         skill_descriptions: {},
         uncategorized_skill_ids: [],
