+++ core/templates/services/UpgradedServices.ts
@@ -238,7 +238,6 @@ import {
   // eslint-disable-next-line max-len
 } from 'interactions/RatioExpressionInput/directives/ratio-expression-input-validation.service';
 import {ReadOnlyCollectionBackendApiService} from 'domain/collection/read-only-collection-backend-api.service';
-import {ReadOnlyTopicObjectFactory} from 'domain/topic_viewer/read-only-topic-object.factory';
 import {ReviewTestBackendApiService} from 'domain/review_test/review-test-backend-api.service';
 import {ReviewTestEngineService} from 'pages/review-test-page/review-test-engine.service';
 import {SchemaDefaultValueService} from 'services/schema-default-value.service';
@@ -937,8 +936,6 @@ export class UpgradedServices {
         upgradedServices['HttpClient'],
         upgradedServices['UrlInterpolationService']
       );
-    upgradedServices['ReadOnlyTopicObjectFactory'] =
-      new ReadOnlyTopicObjectFactory();
     upgradedServices['ReviewTestBackendApiService'] =
       new ReviewTestBackendApiService(
         upgradedServices['HttpClient'],
